Move `CUIRect` class into its own files
