+++ CMakeLists.txt
@@ -2103,6 +2103,8 @@ if(CLIENT)
     ui.h
     ui_ex.cpp
     ui_ex.h
+    ui_rect.cpp
+    ui_rect.h
   )
   set_src(GAME_EDITOR GLOB src/game/editor
     auto_map.cpp
